requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
PyMuPDF==1.23.8
langchain
faiss-cpu
openai